Initial implementation of client form
I had initially been doing the tedious work of creating the scaffolding
for this form myself, but ran into a million other things. I turned to
our friends at Anthropic to help out. Here is the prompt I provided
Codex to get here.

---

We need to set this project up. Do not worry about the server side- at all.

There will be two parts of the client, but we are only working on one today: the input form.

This form will be used to record observations of phenotypic abornmalities present in zebrafish after some sort of toxicological exposure. This form will have several general parts:

  1. Image of observation - The user will add an image of the zebrafish where the observation is being described.

  2. Provenance of data - Information about the submitter and where the data came from.

  3. Information about the fish.

  4. Information about the conditions in which the fish was reared.

  5. Information about the exposure event.

  6. Information about the phenotypic abornmality.

Members of this project have created a requirements document that outlines all of the fields that should be in this form. You can find it in the root directory at `reqs.pdf`.

This project should be built with vite. Do not worry about writing tests just yet. We will be using React to build this form, although we may switch to Preact in the future-- so don't be too fancy: no Suspense, Server-side rendering, or any other recent advanced features in React.

The main entry component, <ZappAtlasApp>, should be rendered in the file `./client/src/index.tsx`. This will be the entry point to the application.

Today you will be working on a component <ZappForm>, which will be the form which implements the requirements in reqs.pdf. <ZappForm> should have a prop, `onChange`, which will look like this:

```
<ZappForm onChange={data => {
    console.log(data);
    // `data` is a JSON object reflecting the data in the form.
}} />
```

Please make a schema for this data using `Zod` in `./client/src/schema.ts`. This schema should be called `ZappObservationSchema` and this module should export a corresponding `ZappObservation` type.

For now, the <ZappAtlasApp> component should just render one `<ZappForm />`. We will do more with it later.

Components should go in `./client/src/components/*.tsx`. Shared UI components (inputs, selects, text boxes) should go in `./client/src/ui/*.tsx`. You should be able to import like `import Input from '@/ui/Input`, or something similar, using `./client/src/` as the root instead of relative paths.

Do not worry about styling this form yet-- we just want to get the components down. Do not add tailwind, or any other styling framework. For the time being, just put things in a basic CSS grid that reflects the structure you find in the PDF. This input form will be used by biologists who are comfortable with editing things in spreadsheets. We do not want to make this a spreadsheet, as there is too much data to fit in a row, but we want to keep that in mind. Try to group things logically.

IMPORTANT: Several of the requirements mention picking terms from an ontology. Mark those fields as TODO for now-- I have a separate UI component for picking ontology elements that I will add myself later.