Add in phenotype picker!
I want to write a longer commit message, but I need to get it in!

Right now, it opens up a modal to select. I don't know if that's best,
but that's how it is. It also sorts by ZFIN usage, but should be able to
sort alphabetically. More to come!